+++ tests/pathogen-embed-pca-multiple-alignments.t
@@ -1,7 +1,7 @@
 Run pathogen-embed with PCA on a H3N2 HA and H3N2 NA alignments.
 
   $ pathogen-embed \
-  >   --alignment $TESTDIR/data/h3n2_ha_alignment.fasta $TESTDIR/data/h3n2_na_alignment.fasta \
+  >   --alignment $TESTDIR/data/h3n2_ha_alignment.sorted.fasta $TESTDIR/data/h3n2_na_alignment.sorted.fasta \
   >   --output-dataframe embed_pca.csv \
   >   pca \
   >   --components 2 \
@@ -9,7 +9,7 @@ Run pathogen-embed with PCA on a H3N2 HA and H3N2 NA alignments.
 
 There should be one record in the embedding per input sequence in the alignment.
 
-  $ [[ $(sed 1d embed_pca.csv | wc -l) == $(grep "^>" $TESTDIR/data/h3n2_ha_alignment.fasta | wc -l) ]]
+  $ [[ $(sed 1d embed_pca.csv | wc -l) == $(grep "^>" $TESTDIR/data/h3n2_ha_alignment.sorted.fasta | wc -l) ]]
 
 There should be 2 components of variance explained.